pyproject.toml
[project]
name = "dbmasta"
version = "0.1.26"
description = "A simple MariaDB/Postgres client based on SQLAlchemy core"
readme = "README.md"
authors = [{ name = "Matt Owen", email = "matt@dealerclear.com" }]
license = { text = "MIT" }
requires-python = ">=3.7"
dependencies = [
    "sqlalchemy>=2.0.41",
    "aiomysql",
    "pymysql",
    "asyncmy",
    "asyncpg",
    "psycopg-binary"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/mastamatto/dbmasta"
Repository = "https://github.com/mastamatto/dbmasta"
Issues = "https://github.com/mastamatto/dbmasta/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
